renamed lcm_rpmsg_bridge to hammer_control_arm and hamemer_control to hammer_control_pru.  Now hammer_control_arm will parse toml config file for configuratioin variables

+++ Turret/TurretControl/inc/hammer_control_pru/hammer_control_pru.h
@@ -21,19 +21,19 @@ extern int32_t g_desired_throw_angle;
 
 // config values, filled in from a CONF message from ARM
 
-extern int32_t g_max_throw_angle;
-extern int32_t g_min_retract_angle;
-extern int32_t g_retract_fill_pressure;
-extern int32_t g_brake_exit_velocity;
-extern int32_t g_emergency_brake_angle;
-extern int32_t g_valve_change_dt;
+extern int64_t g_max_throw_angle;
+extern int64_t g_min_retract_angle;
+extern int64_t g_retract_fill_pressure;
+extern int64_t g_brake_exit_velocity;
+extern int64_t g_emergency_brake_angle;
+extern int64_t g_valve_change_dt;
 
-extern int32_t g_max_throw_pressure_dt;
-extern int32_t g_max_throw_expand_dt;
-extern int32_t g_max_retract_pressure_dt;
-extern int32_t g_max_retract_expand_dt;
-extern int32_t g_max_retract_break_dt;
-extern int32_t g_max_retract_settle_dt;
+extern int64_t g_max_throw_pressure_dt;
+extern int64_t g_max_throw_expand_dt;
+extern int64_t g_max_retract_pressure_dt;
+extern int64_t g_max_retract_expand_dt;
+extern int64_t g_max_retract_break_dt;
+extern int64_t g_max_retract_settle_dt;
 
 // -----------------------------------------------------------------------------
 // public methods
